Clean up implementation of fixtures and tests using Playwright
Move to using new `C3D.Extensions.Playwright.AspNetCore.Xunit` package that encapsulates most of the old PlaywrightWebApplicationFactory.
