+++ src/TagzApp.WebTest/TestsBase.cs
@@ -1,0 +1,21 @@
+﻿using TagzApp.WebTest.Fixtures;
+
+namespace TagzApp.WebTest;
+
+// By default all pages are created in the same browser instance (and may share cookie states etc.) for the scope of the fixture
+// The fixture is scoped to each derived Tests Class.
+// If the tests in a class need to be separated from each other (e.g. testing auth scenarios), then use incognito contexts
+// If required you can launch a new browser instance
+public abstract class TestsBase : IClassFixture<PlaywrightWebApplicationFactory>
+{
+
+	protected readonly PlaywrightWebApplicationFactory WebApp;
+	protected readonly ITestOutputHelper OutputHelper;
+
+	protected TestsBase(PlaywrightWebApplicationFactory webapp, ITestOutputHelper outputHelper)
+	{
+		WebApp = webapp;
+		OutputHelper = outputHelper;
+	}
+
+}
