+++ src/TagzApp.WebTest/xunit.runner.json
@@ -1,0 +1,4 @@
+{
+	"$schema": "https://xunit.net/schema/current/xunit.runner.schema.json",
+	"diagnosticMessages": true
+}
